refactor(solver): split instantiate_key per-family arms into child modules (#13001)
